Fix all compile and link errors blocking bmake kernel on macOS with GCC 16
Resolves every error surfaced by x86_64-elf-gcc -m32 under C23 defaults
so that `bmake kernel` completes cleanly on the feature/macos-build-qemu branch.

Key fixes:
- C23/GCC 16: updated empty () declarations to typed signatures throughout
  syscalls.h, ubixfs.h, ufs.h, atkbd.h (both copies), and ld.h
- stdatomic.h: detect GCC before Clang so GCC 16 uses __GNUC_ATOMICS,
  not the missing __c11_atomic_* builtins
- libcpp.cc: add sized-delete operators for GCC 14+ codegen
- sys/lib/Makefile: add kern_trie.o (was missing, broke sysctl link)
- sys/fs/fat/Makefile: create missing build file for FAT driver objects
- sys/sde/Makefile: use $(MAKE) instead of bare 'make' to stay in bmake
- sys/Makefile.incl: add -Wno-incompatible-pointer-types for syscall tables
- net: forward-declare lwip socket functions instead of including sockets.h
  (whose macros redefine fcntl/close/ioctl and break descrip.h); change
  lnc_netif and tmpBuf in init.c/ethernetif.c to extern
- vfs.c: rename vfsFindFS -> vfs_findFS to match header declaration
- kern_trie.c: fix recursive call (deletion -> delete_trieNode), use kfree
- Numerous missing #include and cast fixes across vmm/, kernel/, pci/, fs/

+++ sys/arch/i386/i386_exec.c
@@ -155,7 +155,7 @@ uint32_t execThread(void (*tproc)(void), uint32_t stack, char *arg) {
   newProcess = schedNewTask();
   assert(newProcess);
 
-  stackAddr = vmm_getFreeKernelPage(newProcess->id, stack / PAGE_SIZE);
+  stackAddr = (uint32_t)vmm_getFreeKernelPage(newProcess->id, stack / PAGE_SIZE);
 
   /* Set All The Correct Thread Attributes */
   newProcess->tss.back_link = 0x0;
